# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Notifuse API",
    "description": "API for Notifuse - a transactional email and newsletter management platform",
    "version": "1.0.0",
    "contact": {
      "name": "Notifuse Support",
      "url": "https://www.notifuse.com/support",
      "email": "hello@notifuse.com"
    }
  },
  "servers": [
    {
      "url": "https://{notifuseDomain}",
      "description": "Customer-specific Notifuse API server",
      "variables": {
        "notifuseDomain": {
          "description": "Your unique Notifuse domain",
          "default": "demo.notifuse.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/contacts.upsert": {
      "post": {
        "summary": "Create or update a contact",
        "description": "Creates a new contact or updates an existing one based on email address. Returns information about whether the contact was created or updated.",
        "operationId": "upsertContact",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact upserted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "workspace_id is required"
                    }
                  },
                  "invalidEmail": {
                    "value": {
                      "error": "invalid email format"
                    }
                  },
                  "missingContact": {
                    "value": {
                      "error": "contact is required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts.getByEmail": {
      "get": {
        "summary": "Get a contact by email address",
        "description": "Retrieves a contact by their email address within a specific workspace. The response always includes the contact's list subscriptions with their status (active, pending, unsubscribed, bounced, complained).",
        "operationId": "getContactByEmail",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "description": "The email address of the contact",
            "example": "user@example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contact": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "Missing workspace ID"
                    }
                  },
                  "missingEmail": {
                    "value": {
                      "error": "Missing email"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Contact not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts.getByExternalID": {
      "get": {
        "summary": "Get a contact by external ID",
        "description": "Retrieves a contact by their external ID within a specific workspace. The response always includes the contact's list subscriptions with their status (active, pending, unsubscribed, bounced, complained).",
        "operationId": "getContactByExternalID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          {
            "name": "external_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The external ID of the contact",
            "example": "user_12345"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contact": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "Missing workspace ID"
                    }
                  },
                  "missingExternalId": {
                    "value": {
                      "error": "Missing external ID"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Contact not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts.import": {
      "post": {
        "summary": "Batch import contacts",
        "description": "Creates or updates multiple contacts in a single batch operation. This is significantly more efficient than individual upsert operations. Optionally subscribes all contacts to specified lists.",
        "operationId": "batchImportContacts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchImportContactsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch import completed (may include partial failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchImportContactsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "workspace_id is required"
                    }
                  },
                  "missingContacts": {
                    "value": {
                      "error": "contacts is required"
                    }
                  },
                  "emptyContacts": {
                    "value": {
                      "error": "contacts array cannot be empty"
                    }
                  },
                  "insufficientPermissions": {
                    "value": {
                      "error": "write access to lists required when subscribe_to_lists is provided"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts.delete": {
      "post": {
        "summary": "Delete a contact",
        "description": "Deletes a contact from the workspace by email address. This action is permanent and cannot be undone.",
        "operationId": "deleteContact",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the deletion was successful",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "workspace_id is required"
                    }
                  },
                  "missingEmail": {
                    "value": {
                      "error": "email is required"
                    }
                  },
                  "invalidEmail": {
                    "value": {
                      "error": "invalid email format"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Contact not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to delete contact"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscribe": {
      "post": {
        "summary": "Subscribe to email lists",
        "description": "Subscribe a contact to one or more email lists. This is a public endpoint that doesn't require authentication. If the list has double opt-in enabled, a confirmation email will be sent.",
        "operationId": "subscribeToLists",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeToListsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the subscription was successful",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingWorkspaceId": {
                    "value": {
                      "error": "workspace_id is required"
                    }
                  },
                  "invalidEmail": {
                    "value": {
                      "error": "invalid contact: invalid email format"
                    }
                  },
                  "missingListIds": {
                    "value": {
                      "error": "list_ids is required"
                    }
                  },
                  "listNotPublic": {
                    "value": {
                      "error": "list is not public"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to subscribe to lists"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token for authentication"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "SubscribeToListsRequest": {
        "type": "object",
        "required": ["workspace_id", "contact", "list_ids"],
        "properties": {
          "workspace_id": {
            "type": "string",
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          "contact": {
            "$ref": "#/components/schemas/SubscriptionContact"
          },
          "list_ids": {
            "type": "array",
            "description": "Array of list IDs to subscribe the contact to",
            "items": {
              "type": "string"
            },
            "example": ["newsletter", "product_updates"],
            "minItems": 1
          }
        }
      },
...
      "UpsertContactRequest": {
        "type": "object",
        "required": ["workspace_id", "contact"],
        "properties": {
          "workspace_id": {
            "type": "string",
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          }
        }
      },
      "UpsertContactResponse": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": ["created", "updated", "unchanged"],
            "description": "The action that was performed on the contact",
            "example": "created"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message if the operation failed",
            "example": null
          }
        }
      },
      "DeleteContactRequest": {
        "type": "object",
        "required": ["workspace_id", "email"],
        "properties": {
          "workspace_id": {
            "type": "string",
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the contact to delete",
            "example": "user@example.com"
          }
        }
      },
      "BatchImportContactsRequest": {
        "type": "object",
        "required": ["workspace_id", "contacts"],
        "properties": {
          "workspace_id": {
            "type": "string",
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          "contacts": {
            "type": "array",
            "description": "Array of contacts to import (create or update)",
            "items": {
              "$ref": "#/components/schemas/Contact"
            },
            "minItems": 1,
            "example": [
              {
                "email": "user1@example.com",
                "first_name": "John",
                "last_name": "Doe"
              },
              {
                "email": "user2@example.com",
                "first_name": "Jane",
                "last_name": "Smith"
              }
            ]
          },
          "subscribe_to_lists": {
            "type": "array",
            "description": "Optional array of list IDs to subscribe all contacts to",
            "items": {
              "type": "string"
            },
            "example": ["newsletter", "product_updates"]
          }
        }
      },
      "BatchImportContactsResponse": {
        "type": "object",
        "properties": {
          "operations": {
            "type": "array",
            "description": "Array of operation results, one for each contact in the request",
            "items": {
              "$ref": "#/components/schemas/UpsertContactOperation"
            }
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Global error message if the entire operation failed",
            "example": null
          }
        }
      },
      "UpsertContactOperation": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the contact",
            "example": "user@example.com"
          },
          "action": {
            "type": "string",
            "enum": ["create", "update", "error"],
            "description": "The action that was performed: 'create' for new contacts, 'update' for existing contacts, 'error' for failed operations",
            "example": "create"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message if this specific contact operation failed",
            "example": null
          }
        }
      },
...
    }
  }
}